# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NestyAI",
    "description": "Personal AI Gateway Server",
    "version": "0.1.0"
  },
  "paths": {
...
    "/internal/console/runtime/providers/openai-compatible": {
      "post": {
        "tags": [
          "internal-console-runtime-providers"
        ],
        "summary": "Create Openai Compatible Provider",
        "operationId": "create_openai_compatible_provider_internal_console_runtime_providers_openai_compatible_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeOpenAIProviderCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Openai Compatible Provider Internal Console Runtime Providers Openai Compatible Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/console/runtime/providers/{provider_id}": {
      "get": {
        "tags": [
          "internal-console-runtime-providers"
        ],
        "summary": "Get Runtime Provider Endpoint",
        "operationId": "get_runtime_provider_endpoint_internal_console_runtime_providers__provider_id__get",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Runtime Provider Endpoint Internal Console Runtime Providers  Provider Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "internal-console-runtime-providers"
        ],
        "summary": "Patch Runtime Provider Endpoint",
        "operationId": "patch_runtime_provider_endpoint_internal_console_runtime_providers__provider_id__patch",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeOpenAIProviderUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Patch Runtime Provider Endpoint Internal Console Runtime Providers  Provider Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "internal-console-runtime-providers"
        ],
        "summary": "Delete Runtime Provider Endpoint",
        "operationId": "delete_runtime_provider_endpoint_internal_console_runtime_providers__provider_id__delete",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Delete Runtime Provider Endpoint Internal Console Runtime Providers  Provider Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/console/runtime/providers/{provider_id}/test": {
      "post": {
        "tags": [
          "internal-console-runtime-providers"
        ],
        "summary": "Test Runtime Provider Endpoint",
        "operationId": "test_runtime_provider_endpoint_internal_console_runtime_providers__provider_id__test_post",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeProviderTestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Test Runtime Provider Endpoint Internal Console Runtime Providers  Provider Id  Test Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/console/runtime/providers/{provider_id}/enable": {
      "post": {
        "tags": [
          "internal-console-runtime-providers"
        ],
        "summary": "Enable Provider Endpoint",
        "operationId": "enable_provider_endpoint_internal_console_runtime_providers__provider_id__enable_post",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Enable Provider Endpoint Internal Console Runtime Providers  Provider Id  Enable Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/console/runtime/providers/{provider_id}/disable": {
      "post": {
        "tags": [
          "internal-console-runtime-providers"
        ],
        "summary": "Disable Provider Endpoint",
        "operationId": "disable_provider_endpoint_internal_console_runtime_providers__provider_id__disable_post",
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Disable Provider Endpoint Internal Console Runtime Providers  Provider Id  Disable Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RuntimeOpenAIProviderCreateRequest": {
        "properties": {
          "provider_id": {
            "type": "string",
            "maxLength": 48,
            "minLength": 3,
            "title": "Provider Id"
          },
          "display_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "base_url": {
            "type": "string",
            "maxLength": 512,
            "minLength": 8,
            "title": "Base Url"
          },
          "chat_completions_path": {
            "type": "string",
            "maxLength": 256,
            "title": "Chat Completions Path",
            "default": "/v1/chat/completions"
          },
          "models_path": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Models Path"
          },
          "api_key_mode": {
            "type": "string",
            "title": "Api Key Mode",
            "default": "none"
          },
          "api_key_env_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Env Name"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4096
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "default_headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Default Headers"
          },
          "default_timeout_seconds": {
            "type": "number",
            "maximum": 600.0,
            "minimum": 1.0,
            "title": "Default Timeout Seconds",
            "default": 30.0
          },
          "supports_streaming": {
            "type": "boolean",
            "title": "Supports Streaming",
            "default": true
          },
          "supports_json_mode": {
            "type": "boolean",
            "title": "Supports Json Mode",
            "default": false
          },
          "supports_tools": {
            "type": "boolean",
            "title": "Supports Tools",
            "default": false
          },
          "supports_reasoning_effort": {
            "type": "boolean",
            "title": "Supports Reasoning Effort",
            "default": false
          },
          "health_check_model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Health Check Model"
          }
        },
        "type": "object",
        "required": [
          "provider_id",
          "display_name",
          "base_url"
        ],
        "title": "RuntimeOpenAIProviderCreateRequest"
      },
      "RuntimeOpenAIProviderUpdateRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "chat_completions_path": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Completions Path"
          },
          "models_path": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Models Path"
          },
          "api_key_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Mode"
          },
          "api_key_env_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Env Name"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4096
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "default_headers": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Headers"
          },
          "default_timeout_seconds": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 600.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Timeout Seconds"
          },
          "supports_streaming": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supports Streaming"
          },
          "supports_json_mode": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supports Json Mode"
          },
          "supports_tools": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supports Tools"
          },
          "supports_reasoning_effort": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supports Reasoning Effort"
          },
          "health_check_model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Health Check Model"
          }
        },
        "type": "object",
        "title": "RuntimeOpenAIProviderUpdateRequest"
      },
...
      "RuntimeProviderTestRequest": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "message": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Message",
            "default": "Reply with exactly: OK"
          }
        },
        "type": "object",
        "title": "RuntimeProviderTestRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}